+++ src/services/Publish.php
@@ -7,8 +7,8 @@
 
 namespace dukt\twitter\services;
 
+use Craft;
 use dukt\twitter\Plugin;
-use GuzzleHttp\Client;
 use yii\base\Component;
 use GuzzleHttp\Exception\GuzzleException;